backend/requirements.txt
# Verdant Backend - Part B
# FastAPI + MongoDB (Motor) + Pydantic v2

# Core
fastapi==0.109.2
uvicorn[standard]==0.27.1
pydantic==2.6.1
pydantic-settings==2.1.0

# MongoDB
motor==3.3.2
pymongo==4.6.1

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Redis (for caching)
redis==5.0.1
aioredis==2.0.1

# Background Workers
celery==5.3.6
rq==1.16.0

# Utilities
python-multipart==0.0.9
python-dateutil==2.8.2
email-validator==2.1.0.post1

# Development
httpx==0.26.0
pytest==8.0.0
pytest-asyncio==0.23.4

# HTTP client (for webhooks)
aiohttp==3.9.3
